+++ pubspec.yaml
@@ -71,7 +71,6 @@ dependencies:
     git:
       url: https://github.com/google/flutter-desktop-embedding.git
       path: plugins/window_size
-  google_fonts: ^6.2.1
 
 dependency_overrides:
   extended_text_field: ^16.0.0
